@@ -2,7 +2,7 @@ package com.mrbattery.widget
 
 import android.content.Context
 import android.graphics.Canvas
-import android.support.constraint.ConstraintLayout
+import androidx.constraintlayout.widget.ConstraintLayout
 import android.util.AttributeSet
 import android.view.View
 import com.mrbattery.R
